+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `XCN` for extended composite names
 - The `CNN` data type now has a `#full_name` method to match `XPN` and `XCN`.
 - The `NDL` data type now has a `#full_name` method to match `XPN` and `XCN`.
+- The segment type now overrides `#to_s`
 
 ### Changed